--- src/main/java/com/markcos/first/todo/Entity/Category.java
@@ -1,22 +1,0 @@
-package com.markcos.first.todo.Entity;
-
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "category")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "icon")
-    private String icon;
-
-    @Column(name = "color")
-    private String color;
-
-}
