+++ src/main/java/com/markcos/first/todo/entity/Category.java
@@ -1,0 +1,24 @@
+package com.markcos.first.todo.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "category")
+@Getter
+@Setter
+public class Category {
+    @Id
+    private Integer id;
+
+    @Column(name = "name")
+    private String name;
+
+    @Column(name = "icon")
+    private String icon;
+
+    @Column(name = "color")
+    private String color;
+
+}
