+++ src/main/java/com/markcos/first/todo/repository/CategoryRepository.java
@@ -1,11 +1,10 @@
 package com.markcos.first.todo.repository;
 
 
-import com.markcos.first.todo.Entity.Category;
+import com.markcos.first.todo.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
