--- src/main/resources/application.properties
@@ -1,18 +1,0 @@
-# DB
-spring.datasource.url=jdbc:postgresql://devhong.asuscomm.com:35432/
-spring.datasource.username=todo
-spring.datasource.password=todo1@#
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-spring.datasource.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-
-# hibernate setting
-spring.jpa.properties.hibernate.use_sql_comments=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.generate-ddl=false
-spring.jpa.show-sql=true
-spring.jpa.database=postgresql
-
