+++ src/main/resources/db/migration/V2__CREATE_CATEGORY.sql
@@ -1,5 +1,5 @@
 create table "category" (
-    id serial constraint categorty_pk primary key,
+    id serial constraint category_pk primary key,
     name varchar(255),
     icon varchar(255),
     color varchar(255)
